Move subnet column to a separate component.
