+++ ui/src/app/machines/views/MachineDetails/MachineNetwork/NetworkTable/NetworkTable.test.tsx
@@ -6,17 +6,13 @@ import NetworkTable from "./NetworkTable";
 
 import { NetworkInterfaceTypes } from "app/store/machine/types";
 import type { RootState } from "app/store/root/types";
-import { NodeStatus } from "app/store/types/node";
 import {
-  networkDiscoveredIP as networkDiscoveredIPFactory,
   fabric as fabricFactory,
   vlan as vlanFactory,
   machineDetails as machineDetailsFactory,
   machineInterface as machineInterfaceFactory,
   machineState as machineStateFactory,
-  networkLink as networkLinkFactory,
   rootState as rootStateFactory,
-  subnet as subnetFactory,
 } from "testing/factories";
 
 const mockStore = configureStore();
@@ -31,6 +27,7 @@ describe("NetworkTable", () => {
       }),
     });
   });
+
   it("displays a spinner when loading", () => {
     state.machine.items = [];
     const store = mockStore(state);
@@ -196,98 +193,5 @@ describe("NetworkTable", () => {
     const links = wrapper.find("DoubleRow[data-test='fabric'] LegacyLink");
     expect(links.at(0).text()).toBe("fabric-name");
     expect(links.at(1).text()).toBe("2 (vlan-name)");
-  });
-
-  it("can display subnet links", () => {
-    const fabric = fabricFactory({ name: "fabric-name" });
-    state.fabric.items = [fabric];
-    const vlan = vlanFactory({ fabric: fabric.id, vid: 2, name: "vlan-name" });
-    const subnet = subnetFactory({ cidr: "subnet-cidr", name: "subnet-name" });
-    state.vlan.items = [vlan];
-    state.subnet.items = [subnet];
-    state.machine.items = [
-      machineDetailsFactory({
-        interfaces: [
-          machineInterfaceFactory({
-            discovered: null,
-            links: [networkLinkFactory({ subnet_id: subnet.id })],
-            vlan_id: vlan.id,
-          }),
-        ],
-        system_id: "abc123",
-      }),
-    ];
-    const store = mockStore(state);
-    const wrapper = mount(
-      <Provider store={store}>
-        <NetworkTable systemId="abc123" />
-      </Provider>
-    );
-    const links = wrapper.find("DoubleRow[data-test='subnet'] LegacyLink");
-    expect(links.at(0).text()).toBe("subnet-cidr");
-    expect(links.at(1).text()).toBe("subnet-name");
-  });
-
-  it("can display an unconfigured subnet", () => {
-    const fabric = fabricFactory({ name: "fabric-name" });
-    state.fabric.items = [fabric];
-    const vlan = vlanFactory({ fabric: fabric.id, vid: 2, name: "vlan-name" });
-    state.vlan.items = [vlan];
-    state.subnet.items = [];
-    state.machine.items = [
-      machineDetailsFactory({
-        interfaces: [
-          machineInterfaceFactory({
-            discovered: null,
-            links: [networkLinkFactory({ subnet_id: 99 })],
-            vlan_id: vlan.id,
-          }),
-        ],
-        system_id: "abc123",
-      }),
-    ];
-    const store = mockStore(state);
-    const wrapper = mount(
-      <Provider store={store}>
-        <NetworkTable systemId="abc123" />
-      </Provider>
-    );
-    expect(wrapper.find("DoubleRow[data-test='subnet']").prop("primary")).toBe(
-      "Unconfigured"
-    );
-  });
-
-  it("can display the subnet name only", () => {
-    const fabric = fabricFactory({ name: "fabric-name" });
-    state.fabric.items = [fabric];
-    const vlan = vlanFactory({ fabric: fabric.id, vid: 2, name: "vlan-name" });
-    const subnet = subnetFactory({ cidr: "subnet-cidr", name: "subnet-name" });
-    state.vlan.items = [vlan];
-    state.subnet.items = [subnet];
-    state.machine.items = [
-      machineDetailsFactory({
-        interfaces: [
-          machineInterfaceFactory({
-            discovered: [networkDiscoveredIPFactory({ subnet_id: subnet.id })],
-            links: [],
-            vlan_id: vlan.id,
-          }),
-        ],
-        status: NodeStatus.DEPLOYING,
-        system_id: "abc123",
-      }),
-    ];
-    const store = mockStore(state);
-    const wrapper = mount(
-      <Provider store={store}>
-        <NetworkTable systemId="abc123" />
-      </Provider>
-    );
-    expect(
-      wrapper.find("DoubleRow[data-test='subnet'] LegacyLink").exists()
-    ).toBe(false);
-    expect(wrapper.find("DoubleRow[data-test='subnet']").prop("primary")).toBe(
-      "subnet-cidr (subnet-name)"
-    );
   });
 });
