+++ ui/src/app/machines/views/MachineDetails/MachineNetwork/NetworkTable/SubnetColumn/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./SubnetColumn";
